added new figure - rectangle

+++ figures/include/Triangle.h
@@ -4,13 +4,13 @@
 
 class Triangle : public Figure
 {
-	int a;
-	int b;
-	int c;
+	double a;
+	double b;
+	double c;
 
 	void setStrRepresentation();
 public:
-	Triangle(int a, int b, int c);
+	Triangle(double a, double b, double c);
 	double getPerimeter() const override;
 	std::string toString() const override;
 	std::unique_ptr<Figure> clone() const override;
